packages are being dropped in drop circles some problems with identifying what is a tree and what is a drop point when get close within 20-30 m of a lidar sample. It seems I am often identifying a tree in peripheral "vision" as drop point and then making a delivery
avoiding trees was working, i made it try to avoid faster in order to not
have such problems with the edge cases, now I'm crashing straight into them

i decided to save prior tree list in order to not have trees switch to drops
or vice versa when get close. I'd like to find a better way to get around this.
It also isn't that robust yet anyways.